Minor changes to test + release notes + docs
