+++ tests/framework/cli/pipeline/conftest.py
@@ -1,8 +1,22 @@
+import json
 import shutil
+from pathlib import Path
 
 import pytest
 
 from kedro.framework.project import settings
+
+
+def _write_json(filepath: Path, content: dict):
+    filepath.parent.mkdir(parents=True, exist_ok=True)
+    json_str = json.dumps(content, indent=4)
+    filepath.write_text(json_str)
+
+
+def _write_dummy_file(filepath: Path, content: str = ""):
+    filepath.parent.mkdir(parents=True, exist_ok=True)
+    with filepath.open("w") as f:
+        f.write(content)
 
 
 @pytest.fixture(autouse=True)
@@ -82,3 +96,39 @@ def cleanup_pyproject_toml(fake_repo_path):
     yield
 
     pyproject_toml.write_text(existing_toml)
+
+
+@pytest.fixture()
+def fake_local_template_dir(fake_repo_path):
+    """Set up a local template directory. This won't be functional we're just testing the actual layout works.
+
+    Note that this is not scoped to module because we don't want to have this folder present in most of the tests,
+    so we will tear it down every time.
+    """
+    template_path = fake_repo_path / Path("templates")
+    pipeline_template_path = template_path / Path("pipeline")
+    cookiecutter_template_path = (
+        pipeline_template_path / "{{ cookiecutter.pipeline_name }}"
+    )
+
+    cookiecutter_template_path.mkdir(parents=True)
+
+    # Create the absolute bare minimum files
+    cookiecutter_json = {
+        "pipeline_name": "default",
+    }
+    _write_json(pipeline_template_path / "cookiecutter.json", cookiecutter_json)
+    _write_dummy_file(
+        cookiecutter_template_path / "pipeline_{{ cookiecutter.pipeline_name }}.py",
+    )
+    _write_dummy_file(cookiecutter_template_path / "__init__.py", "")
+    _write_dummy_file(
+        cookiecutter_template_path
+        / r"config/parameters/{{ cookiecutter.pipeline_name }}.yml",
+    )
+    _write_dummy_file(
+        cookiecutter_template_path / r"tests/test_{{ cookiecutter.pipeline_name }}.py",
+    )
+    yield template_path.resolve()
+
+    shutil.rmtree(template_path)
